Add files via upload
The Task is to convert the colored image into a pencil sketch. CV2 library is used to process the image into a sketch of the image using python programming. Since Images are arrays, we can process the images into grayscale and inverted blurry images, and later by applying the divide function, the sketch can be generated.