Add files via upload
The Project is about predicting the type of flower, using a Decision Tree Classifier (Machine Learning algorithm). And when new data is given to the classifier, would it be able to predict the right species of Iris accordingly?

The data set contains 3 classes of 50 instances each, where each class refers to a type of iris plant.